Manage cart of different users

+++ src/db/cart.router.js
@@ -4,19 +4,29 @@ const router = express.Router();
 
 router.use(express.json());
 
-// http://localhost:4000/cart
-router.get('/', async (req, res) => {
+// http://localhost:4000/userId/cart
+router.get('/:userId/cart', async (req, res, next) => {
     try {
-        const cart = await Cart.find();
-        res.send(cart);
+        const { userId } = req.params;
+        Cart.find({user_id: userId}, async function (err, cart) {
+            if (err) {
+                return next(new Error(err));
+            }
+        
+            if (!cart) {
+                return next(new Error('No items in cart'));
+            }
+            console.log(cart);
+            res.send(cart);
+        });
     } catch (error) {
         res.status(500).send(error);
     }
   });
 
 // http://localhost:4000/cart
-router.post('/', async (req, res) => {
-    try {
+router.post('/cart', async (req, res) => {
+    try { 
         const cart = new Cart(req.body);
         await cart.save();
         res.status(201).send(cart);
@@ -25,8 +35,8 @@ router.post('/', async (req, res) => {
     }
 });
 
-// http://localhost:4000/cart/id
-router.delete('/:id', async (req, res) => {
+// http://localhost:4000/userId/cart/id
+router.delete('/cart/:id', async (req, res) => {
     try {
         const { id } = req.params;
         const cart = await Cart.findById(id);
@@ -37,10 +47,11 @@ router.delete('/:id', async (req, res) => {
     }
 });
 
-// http://localhost:4000/cart
-router.delete('/', async (req, res) => {
+// http://localhost:4000/userId/cart
+router.delete('/:userId/cart', async (req, res) => {
     try {
-        await Cart.deleteMany({});
+        const { userId } = req.params;
+        await Cart.deleteMany({user_id: userId});
         res.send("Cart has been cancelled");
     } catch (error) {
         res.status(500).send(error);
